+++ src/grameneSwaggerClient.js
@@ -1,0 +1,15 @@
+'use strict';
+
+var _ = require('lodash');
+var Q = require('q');
+
+var grameneClientPromise = (function() {
+  var Client = require('swagger-client');
+  var deferred = Q.defer();
+  var gramene = new Client({url: 'http://devdata.gramene.org/swagger', success: function() {
+    deferred.resolve(gramene);
+  }});
+  return deferred.promise;
+})();
+
+module.exports = grameneClientPromise;
